// admin-dashboard/src/app/services/models/user-profile.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserProfile {
  createdAt?: string;
  firstName: string;
  lastName: string;
  nationalId?: string;
  phone: string;
  phoneVerified?: boolean;
  profilePic?: string;
  publicId: string;
  remark?: string;
  roles: Array<'SUPER_ADMIN' | 'ADMIN' | 'DRIVER' | 'CONSIGNOR'>;
  status: 'OTP_SENT' | 'VERIFIED' | 'PENDING' | 'APPROVED' | 'ACTIVE' | 'DISABLED' | 'SUSPENDED' | 'REJECTED';
  updatedAt?: string;
}